requirements.txt
python-telegram-bot
gTTS
SpeechRecognition
lyricsgenius
pytest-shutil
Wave
pydub
ffmpeg
giphypop
